+++ Questionable/Controller/Steps/Common/SendNotification.cs
@@ -41,6 +41,7 @@ internal static class SendNotification
 
     internal sealed class Executor
     (
+        NotificationMasterIpc notificationMasterIpc,
         IChatGui chatGui,
         Configuration configuration) : TaskExecutor<Task>
     {
@@ -100,7 +101,7 @@ internal static class SendNotification
                 };
                 chatGui.Print(message);
             }
-
+            notificationMasterIpc.Notify(text);
             return true;
         }