+++ Questionable/Controller/Steps/TaskCreator.cs
@@ -19,6 +19,7 @@ internal sealed class TaskCreator
     TerritoryData territoryData,
     IClientState clientState,
     IChatGui chatGui,
+    Configuration configuration,
     ILogger<TaskCreator> logger)
 {
     private readonly IChatGui _chatGui = chatGui;
@@ -30,8 +31,8 @@ internal sealed class TaskCreator
     public IReadOnlyList<ITask> CreateTasks(Quest quest, byte sequenceNumber, QuestSequence? sequence, QuestStep? step)
     {
         List<ITask> newTasks;
-# if !DEBUG
-        if (quest.Root.Disabled && sequenceNumber.InRange(1, 2))
+
+        if (!configuration.Advanced.Debug && quest.Root.Disabled && sequenceNumber.InRange(1, 2))
         {
             var reason = (quest.Root.Comment ?? "<no reason specified>").Split('\n', 2)[0];
             _chatGui.PrintError(_LF("The quest '{0}' has been marked as Disabled for the following reason: {1}", quest.Info.Name, reason),
@@ -41,7 +42,7 @@ internal sealed class TaskCreator
             _chatGui.PrintError(_L("Thank you for your patience as we expand QST's support to include this quest in a future update."),
                 CommandHandler.MessageTag, CommandHandler.TagColor);
         }
-# endif
+
         if (sequence == null)
         {
             if (!quest.Root.Disabled)